separacao do main
