+++ data/compras.json
@@ -13,9 +13,12 @@
                     123.0
                 ],
                 "roupa": [
-                    40.0
+                    40.0,
+                    2.0
                 ],
-                "lazer": []
+                "lazer": [
+                    90.0
+                ]
             }
         ]
     },
